mlops-experiments/experiments/requirements.txt
mlflow==2.11.3
scikit-learn==1.4.2
python-dotenv==1.0.1
boto3==1.34.69
prometheus-client==0.19.0
numpy==1.26.4
pandas==2.2.2
setuptools==69.5.1